+++ Core/World/texl.cs
@@ -25,7 +25,7 @@ namespace OpenVIII
                     var timOffset = i * TEX_SIZE;
                     var tim = new TIM2(texlBuffer, (uint)timOffset);
                     textures[i] = new TextureHandler[tim.GetClutCount];
-                    for (ushort k = 0; k < textures[i].Length; k++)
+                    for (byte k = 0; k < textures[i].Length; k++)
                         textures[i][k] = TextureHandler.Create($"texl_tim{(i + 1).ToString("D2")}.tim", tim, k, null);
                     //todo detect if mods aren't using palettes.
                 }
